Implemented the hull function.
This may seem like quite a coarse-grained commit, but the implementation
is essentially a copy of the prototype that was used to paint the hulls
that could then be visually inspected.

The purpose of this test suite is actually to enable refactoring of the
hull function.

The main problem of the hull function is the procedural hullPoints
function, which relies on mutation.